[#0] TF : return loaded model from save, and adding hasId() for all models as the check for whether record exists or not. Next to implement is updates

+++ androidsql/src/main/java/com/abstractx1/androidsql/factories/SqlQueryFactory.java
@@ -59,7 +59,7 @@ public class SqlQueryFactory extends Factory {
         return query;
     }
 
-    public static String buildFindById(String tableName, int id) {
+    public static String buildFindById(String tableName, long id) {
         String query = String.format("SELECT * FROM " + tableName + " WHERE id = %d", id);
         return query;
     }
